Move goal types to src/goals
