+++ src/components/GoalTimeline/DefaultState.tsx
@@ -2,14 +2,14 @@ import React from "react";
 import { Goal, GoalsState } from "../../types/goals";
 import Stack from "../../types/stack";
 import MergeDupStep from "../../goals/MergeDupGoal/MergeDupStep";
-import { MergeDups } from "../../types/MergeDups";
-import { CreateCharInv } from "../../types/CreateCharInv";
-import { CreateStrWordInv } from "../../types/CreateStrWordInv";
-import { HandleFlags } from "../../types/HandleFlags";
-import { SpellCheckGloss } from "../../types/SpellCheckGloss";
-import { ValidateChars } from "../../types/ValidateChars";
-import { ValidateStrWords } from "../../types/ValidateStrWords";
-import { ViewFinal } from "../../types/ViewFinal";
+import { MergeDups } from "../../goals/MergeDups/MergeDups";
+import { CreateCharInv } from "../../goals/CreateCharInv/CreateCharInv";
+import { CreateStrWordInv } from "../../goals/CreateStrWordInv/CreateStrWordInv";
+import { HandleFlags } from "../../goals/HandleFlags/HandleFlags";
+import { SpellCheckGloss } from "../../goals/SpellCheckGloss/SpellCheckGloss";
+import { ValidateChars } from "../../goals/ValidateChars/ValidateChars";
+import { ValidateStrWords } from "../../goals/ValidateStrWords/ValidateStrWords";
+import { ViewFinal } from "../../goals/ViewFinal/ViewFinal";
 
 let goal1: Goal = new CreateCharInv([]);
 let goal2: Goal = new CreateStrWordInv([]);
