+++ src/components/GoalTimeline/tests/GoalTimelineActions.test.tsx
@@ -1,19 +1,12 @@
 import * as actions from "../GoalTimelineActions";
 import { Goal } from "../../../types/goals";
-import { BaseGoal } from "../../../types/baseGoal";
-import { User } from "../../../types/User";
 import configureMockStore from "redux-mock-store";
 import { defaultState } from "../DefaultState";
 import thunk from "redux-thunk";
+import { CreateCharInv } from "../../../goals/CreateCharInv/CreateCharInv";
 
 it("should create an action to add a goal", () => {
-  const user: User = {
-    name: "Test user",
-    username: "Test username",
-    id: 0
-  };
-  const goal: Goal = new BaseGoal();
-  goal.user = user;
+  const goal: Goal = new CreateCharInv([]);
   const expectedAction = {
     type: actions.ADD_GOAL,
     payload: goal
@@ -29,14 +22,7 @@ it("should create an async action to add a goal", () => {
       ...defaultState
     }
   });
-
-  const user: User = {
-    name: "Test user",
-    username: "Test username",
-    id: 0
-  };
-  const goal: Goal = new BaseGoal();
-  goal.user = user;
+  const goal: Goal = new CreateCharInv([]);
   const expectedAction = {
     type: actions.ADD_GOAL,
     payload: goal
