+++ src/goals/CreateCharInv/CreateCharInv.tsx
@@ -1,10 +1,10 @@
 import React from "react";
-import { Goal, GoalData, Tools, GoalOption } from "./goals";
-import { User } from "./User";
-import BaseGoalScreen from "../goals/DefaultGoal/BaseGoalScreen/BaseGoalScreen";
-import BaseGoalSelect from "../goals/DefaultGoal/BaseGoalSelect/BaseGoalSelect";
+import { Goal, GoalData, Tools, GoalOption } from "../../types/goals";
+import { User } from "../../types/User";
+import BaseGoalScreen from "../DefaultGoal/BaseGoalScreen/BaseGoalScreen";
+import BaseGoalSelect from "../DefaultGoal/BaseGoalSelect/BaseGoalSelect";
 
-export class CreateStrWordInv implements Goal {
+export class CreateCharInv implements Goal {
   id: number;
   name: string;
   user: User;
@@ -19,7 +19,7 @@ export class CreateStrWordInv implements Goal {
 
   constructor(steps: JSX.Element[]) {
     this.id = 0;
-    this.name = "createStrWordInv";
+    this.name = "createCharInv";
     this.user = new User("", "", 1);
     this.display = <BaseGoalScreen goal={this} />;
     this.goalWidget = <BaseGoalSelect goal={this} />;
