+++ src/types/action.tsx
@@ -1,6 +1,6 @@
 import { Action } from "redux";
 import { Goal } from "./goals";
-import { BaseGoal } from "./baseGoal";
+import { CreateCharInv } from "../goals/CreateCharInv/CreateCharInv";
 
 export interface ActionWithPayload<T> extends Action {
   payload: T;
@@ -14,7 +14,7 @@ interface MockAction extends ActionWithPayload<Goal> {
   payload: Goal;
 }
 
-const goal: Goal = new BaseGoal();
+const goal: Goal = new CreateCharInv([]);
 
 // Used for passing a non-existent action to a reducer
 export const MockActionInstance: MockAction = {
